expert-dispatch: send request activations as f32, completing the f16 removal
The 2026-08-04 f16 fix repaired only the RETURN direction (pipe_expert_partial).
The REQUEST direction stayed f16, so every routed expert RECEIVED f16(x) -- about
3e-4 relative -- while attention, the shared expert and the residual all saw f32
x. An asymmetric, deterministic perturbation on the input of every routed expert
on every layer.

Found independently by both external reviewers (Codex gpt-5.6-sol and Kimi K3)
while reviewing the SwiGLU clamp bug.

WHY NO PROBE COULD SEE IT: WP_SELFCHECK compares gather against dense inside a
worker, and BOTH arms consume the same already-rounded activation. An equivalence
probe proves only that two paths agree; when both share the defect it reports
"ok" forever. WP_EXPERT_GATHER=0 is likewise not "dense everywhere" -- it only
selects the dense-masked graph inside a worker, leaving the f16 input untouched.

  pipe_expert_dispatch_req::activations  vector<uint16_t> -> vector<float>
  encode/decode 2 bytes/value -> 4, activation length check updated
  spine: ggml_fp32_to_fp16_row -> a straight memcpy (contiguous) or f32 copy
  worker: the fp16->fp32 widening loop is gone; it recovered the storage type but
          never the precision the spine had already discarded
  PIPE_VERSION 3 -> 4, so a stale binary REFUSES rather than mis-decoding a
          2-byte stream as 4-byte values

Costs 2x bytes on the request path. Decode is page-in bound (~1 GB of expert
weights per generated token), so this is expected to be invisible -- but that is
a measurement, not an assumption, and it has not been made yet.

ALSO FIXES A TEST THAT HAD SILENTLY GONE VACUOUS. test-wp-expert-worker verified
returned partials with

    ggml_fp16_to_fp32((ggml_fp16_t) response.partial[i])

but partial became std::vector<float> in PIPE_VERSION 2. ggml_fp16_t is uint16_t,
so the float was TRUNCATED to an integer before reinterpretation: 0.35f came back
as exactly 0.0f (verified empirically against ggml). The assertion could only
pass while every expected value sat within tolerance of zero. It now reads the
float directly, and the suite passes with the check actually live -- which also
confirms the f32 request path end-to-end.

Both suites rebuilt. test-wp-expert-worker: all tests passed.
test-wp-expert-dispatcher: unchanged pre-existing marginal-cost speed-split
failure only (see the task filed for it); its swiglu-clamp round-trip assertion
passes.

+++ src/pipeline/pipe-expert-dispatcher.cpp
@@ -682,7 +682,7 @@ struct dispatcher::impl {
 
     std::vector<planned_request> plan_requests(int32_t                                     layer,
                                                uint32_t                                    n_tokens,
-                                               const std::vector<uint16_t> &               activations,
+                                               const std::vector<float> &                  activations,
                                                const std::vector<pipe_expert_assignment> & assignments,
                                                const std::vector<std::vector<size_t>> &   layer_routes,
                                                std::vector<size_t> &                       assigned_counts,
@@ -965,7 +965,7 @@ struct dispatcher::impl {
     std::vector<float> dispatch(int32_t                                     layer,
                                 uint64_t                                    seq_id,
                                 uint32_t                                    n_tokens,
-                                const std::vector<uint16_t> &               activations,
+                                const std::vector<float> &                  activations,
                                 const std::vector<pipe_expert_assignment> & assignments,
                                 float                                       swiglu_clamp) {
         if (poisoned) {
@@ -1243,7 +1243,7 @@ dispatcher & dispatcher::operator=(dispatcher &&) noexcept = default;
 std::vector<float> dispatcher::dispatch(int32_t                                     layer,
                                         uint64_t                                    seq_id,
                                         uint32_t                                    n_tokens,
-                                        const std::vector<uint16_t> &               activations,
+                                        const std::vector<float> &                  activations,
                                         const std::vector<pipe_expert_assignment> & assignments,
                                         float                                       swiglu_clamp) {
     return pimpl->dispatch(layer, seq_id, n_tokens, activations, assignments, swiglu_clamp);
